+++ src/main/java/com/pixelandtag/sms/mt/workerthreads/HttpBillingWorker.java
@@ -466,22 +466,22 @@ public class HttpBillingWorker implements Runnable {
 					
 					if(billable.isSuccess() ||  "Success".equals(billable.getResp_status_code()) ){
 						cmp_ejb.updateMessageInQueue(billable.getCp_tx_id(),BillingStatus.SUCCESSFULLY_BILLED);
-						cmp_ejb.updateSMSStatLog(BigInteger.valueOf(billable.getCp_tx_id()),ERROR.Success);
+						cmp_ejb.updateSMSStatLog(billable.getCp_tx_id(),ERROR.Success);
 						billable.setResp_status_code(BillingStatus.SUCCESSFULLY_BILLED.toString());
 					}
 					if("TWSS_101".equals(billable.getResp_status_code()) || "TWSS_114".equals(billable.getResp_status_code()) || "TWSS_101".equals(billable.getResp_status_code())){
 						cmp_ejb.updateMessageInQueue(billable.getCp_tx_id(),BillingStatus.BILLING_FAILED_PERMANENTLY);
-						cmp_ejb.updateSMSStatLog(BigInteger.valueOf(billable.getCp_tx_id()),ERROR.InvalidSubscriber);
+						cmp_ejb.updateSMSStatLog(billable.getCp_tx_id(),ERROR.InvalidSubscriber);
 						billable.setResp_status_code(BillingStatus.BILLING_FAILED_PERMANENTLY.toString());
 					}
 					if("OL402".equals(billable.getResp_status_code()) || "OL404".equals(billable.getResp_status_code()) || "OL405".equals(billable.getResp_status_code())  || "OL406".equals(billable.getResp_status_code())){
 						cmp_ejb.updateMessageInQueue(billable.getCp_tx_id(),BillingStatus.INSUFFICIENT_FUNDS);
-						cmp_ejb.updateSMSStatLog(BigInteger.valueOf(billable.getCp_tx_id()),ERROR.PSAInsufficientBalance);
+						cmp_ejb.updateSMSStatLog(billable.getCp_tx_id(),ERROR.PSAInsufficientBalance);
 						billable.setResp_status_code(BillingStatus.INSUFFICIENT_FUNDS.toString());
 					}
 					
 					if("TWSS_109".equals(billable.getResp_status_code())){
-						cmp_ejb.updateSMSStatLog(BigInteger.valueOf(billable.getCp_tx_id()),ERROR.PSAChargeFailure);
+						cmp_ejb.updateSMSStatLog(billable.getCp_tx_id(),ERROR.PSAChargeFailure);
 						billable.setIn_outgoing_queue(0L);
 						billable.setProcessed(0L);
 						billable.setRetry_count( (billable.getRetry_count()+1 ) );
