+++ src/main/java/com/pixelandtag/staticcontent/ContentRetriever.java
@@ -1,6 +1,7 @@
 package com.pixelandtag.staticcontent;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
@@ -372,7 +373,7 @@ public class ContentRetriever {
 				
 				ServiceProcessorI processor =  MOProcessorFactory.getProcessorClass(procDTO.getProcessorClassName(), GenericServiceProcessor.class);
 				mo = new MOSms();
-				mo.setCMP_Txid(SubscriptionMain.generateNextTxId());
+				mo.setCMP_Txid(BigInteger.valueOf(SubscriptionMain.generateNextTxId()));
 				mo.setMsisdn(msisdn);
 				mo.setCMP_AKeyword(sm.getCmp_keyword());
 				mo.setCMP_SKeyword(sm.getCmp_skeyword());
